pyproject.toml
[project]
name = "my_project"
version = "0.1.0"
description = "Clustering of molecular data using KMeans and MaxMin algorithms"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.8"

dependencies = [
    "numpy",
    "pandas",
    "rdkit",
    "ruff",
    "scikit-learn",
    "umap-learn"
]

[project.scripts]
run-clustering = "run_clustering:main"
